patch torus precision

+++ raytracerframework_cpp/deg4solver.cpp
@@ -32,9 +32,9 @@
 #include<cmath>
 #include"deg4solver.h"
 
-int SolveQuadric(double c[ 3 ],double s[ 2 ])
+int SolveQuadric(long double c[ 3 ],long double s[ 2 ])
 {
-    double p, q, D;
+    long double p, q, D;
 
     /* normal form: x^2 + px + q = 0 */
 
@@ -54,7 +54,7 @@ int SolveQuadric(double c[ 3 ],double s[ 2 ])
     }
     else /* if (D > 0) */
     {
-	double sqrt_D = sqrt(D);
+	long double sqrt_D = sqrt(D);
 
 	s[ 0 ] =   sqrt_D - p;
 	s[ 1 ] = - sqrt_D - p;
@@ -63,13 +63,13 @@ int SolveQuadric(double c[ 3 ],double s[ 2 ])
 }
 
 
-int SolveCubic(double c [ 4 ],double s[ 3 ])
+int SolveCubic(long double c [ 4 ],long double s[ 3 ])
 {
     int     i, num;
-    double  sub;
-    double  A, B, C;
-    double  sq_A, p, q;
-    double  cb_p, D;
+    long double  sub;
+    long double  A, B, C;
+    long double  sq_A, p, q;
+    long double  cb_p, D;
 
     /* normal form: x^3 + Ax^2 + Bx + C = 0 */
 
@@ -96,9 +96,9 @@ int SolveCubic(double c [ 4 ],double s[ 3 ])
 	    s[ 0 ] = 0;
 	    num = 1;
 	}
-	else /* one single and one double solution */
+	else /* one single and one long double solution */
 	{
-	    double u = cbrt(-q);
+	    long double u = cbrt(-q);
 	    s[ 0 ] = 2 * u;
 	    s[ 1 ] = - u;
 	    num = 2;
@@ -106,8 +106,8 @@ int SolveCubic(double c [ 4 ],double s[ 3 ])
     }
     else if (D < 0) /* Casus irreducibilis: three real solutions */
     {
-	double phi = 1.0/3 * acos(-q / sqrt(-cb_p));
-	double t = 2 * sqrt(-p);
+	long double phi = 1.0/3 * acos(-q / sqrt(-cb_p));
+	long double t = 2 * sqrt(-p);
 
 	s[ 0 ] =   t * cos(phi);
 	s[ 1 ] = - t * cos(phi + M_PI / 3);
@@ -116,9 +116,9 @@ int SolveCubic(double c [ 4 ],double s[ 3 ])
     }
     else /* one real solution */
     {
-	double sqrt_D = sqrt(D);
-	double u = cbrt(sqrt_D - q);
-	double v = - cbrt(sqrt_D + q);
+	long double sqrt_D = sqrt(D);
+	long double u = cbrt(sqrt_D - q);
+	long double v = - cbrt(sqrt_D + q);
 
 	s[ 0 ] = u + v;
 	num = 1;
@@ -135,12 +135,12 @@ int SolveCubic(double c [ 4 ],double s[ 3 ])
 }
 
 
-int SolveQuartic(double c[ 5 ],double s[ 4 ])
+int SolveQuartic(long double c[ 5 ],long double s[ 4 ])
 {
-    double  coeffs[ 4 ];
-    double  z, u, v, sub;
-    double  A, B, C, D;
-    double  sq_A, p, q, r;
+    long double  coeffs[ 4 ];
+    long double  z, u, v, sub;
+    long double  A, B, C, D;
+    long double  sq_A, p, q, r;
     int     i, num;
 
     /* normal form: x^4 + Ax^3 + Bx^2 + Cx + D = 0 */
